+++ lib/persistence.dart
@@ -11,6 +11,7 @@ export 'src/persistence/dao/dao_type.dart';
 export 'src/persistence/query/expression.dart';
 export 'src/persistence/query/filter.dart';
 export 'src/persistence/query/sort.dart';
+export 'src/persistence/query/query_result.dart';
 export 'src/persistence/query/query_select.dart';
 export 'src/persistence/query/query_source.dart';