+++ lib/src/persistence/database_context.dart
@@ -6,22 +6,24 @@ import 'package:datahub/datahub.dart';
 /// via a [DatabaseConnection].
 abstract class DatabaseContext {
   Future<List<TDao>> query<TDao>(
-    DaoDataBean<TDao> bean, {
+    QuerySource<TDao> bean, {
     Filter filter = Filter.empty,
+    List<QuerySelect> distinct = const <QuerySelect>[],
     Sort sort = Sort.empty,
     int offset = 0,
     int limit = -1,
   });
 
   Future<TDao?> queryId<TDao, TPrimaryKey>(
-      PKDaoDataBean<TDao, TPrimaryKey> bean, TPrimaryKey id);
+      PrimaryKeyDataBean<TDao, TPrimaryKey> bean, TPrimaryKey id);
 
   Future<bool> idExists<TPrimaryKey>(
-      PrimaryKeyDataBean<TPrimaryKey> bean, TPrimaryKey id);
+      PrimaryKeyDataBean<dynamic, TPrimaryKey> bean, TPrimaryKey id);
 
   Future<List<dynamic>> select(
     QuerySource bean,
     List<QuerySelect> select, {
+    List<QuerySelect> distinct = const <QuerySelect>[],
     Filter filter = Filter.empty,
     Sort sort = Sort.empty,
     int offset = 0,
@@ -31,20 +33,28 @@ abstract class DatabaseContext {
   /// Returns primary key of inserted object.
   Future<dynamic> insert<TDao extends BaseDao>(TDao object);
 
-  Future<void> update<TDao extends PKBaseDao>(TDao object);
+  Future<void> update<TDao extends PrimaryKeyDao>(TDao object);
 
-  Future<void> updateId<TPrimaryKey>(PrimaryKeyDataBean<TPrimaryKey> bean,
-      TPrimaryKey id, Map<String, dynamic> values);
+  Future<void> updateId<TPrimaryKey>(
+    PrimaryKeyDataBean<dynamic, TPrimaryKey> bean,
+    TPrimaryKey id,
+    Map<String, dynamic> values,
+  );
 
   /// Returns number of affected rows.
   Future<int> updateWhere(
-      BaseDataBean bean, Map<String, dynamic> values, Filter filter);
+    QuerySource source,
+    Map<String, dynamic> values,
+    Filter filter,
+  );
 
-  Future<void> delete<TDao extends PKBaseDao>(TDao object);
+  Future<void> delete<TDao extends PrimaryKeyDao>(TDao object);
 
   Future<void> deleteId<TPrimaryKey>(
-      PrimaryKeyDataBean<TPrimaryKey> bean, TPrimaryKey id);
+    PrimaryKeyDataBean<dynamic, TPrimaryKey> bean,
+    TPrimaryKey id,
+  );
 
   /// Returns number of affected rows.
-  Future<int> deleteWhere(DaoDataBean bean, Filter filter);
+  Future<int> deleteWhere(DataBean bean, Filter filter);
 }