+++ lib/src/persistence/query/query_result.dart
@@ -1,0 +1,6 @@
+class QueryResult {
+  final String layoutName;
+  final Map<String, dynamic> data;
+
+  QueryResult(this.layoutName, this.data);
+}