+++ lib/src/persistence/query/query_source.dart
@@ -1,13 +1,19 @@
+import 'package:boost/boost.dart';
 import 'package:datahub/persistence.dart';
+import 'package:datahub/src/persistence/query/query_result.dart';
 
 /// Defining a common base class among query sources.
 ///
 /// A QuerySource is either a [BaseDataBean] or a [JoinedQuerySource]
-abstract class QuerySource {}
+abstract class QuerySource<T> {
+  const QuerySource();
+
+  T map(List<QueryResult> results);
+}
 
 //TODO use filter objects for join
-class BeanJoin {
-  final BaseDataBean bean;
+class BeanJoin<TDao> {
+  final DataBean<TDao> bean;
   final DataField mainField;
   final CompareType type;
   final DataField beanField;
@@ -15,21 +21,55 @@ class BeanJoin {
   BeanJoin(this.bean, this.mainField, this.type, this.beanField);
 }
 
-class JoinedQuerySource extends QuerySource {
-  final BaseDataBean main;
-  final List<BeanJoin> joins;
+abstract class JoinedQuerySource {
+  DataBean get main;
 
-  JoinedQuerySource(this.main, this.joins);
+  List<BeanJoin> get joins;
+}
+
+class TupleJoinQuerySource<Ta, Tb> extends QuerySource<Tuple<Ta, Tb>> {
+  final DataBean<Ta> main;
+  final BeanJoin<Tb> joinB;
+
+  List<BeanJoin> get joins => [joinB];
+
+  TupleJoinQuerySource(this.main, this.joinB);
+
+  @override
+  Tuple<Ta, Tb> map(List<QueryResult> results) {
+    return Tuple(main.map(results), joinB.bean.map(results));
+  }
 
   /// Create [JoinedQuerySource] with this and the following join.
   ///
   /// This enabled chaining of join() calls.
-  JoinedQuerySource join(
-      BaseDataBean other, DataField mainField, DataField otherField,
+  TripleJoinQuerySource<Ta, Tb, TDao> join<TDao>(
+      DataBean<TDao> other, DataField mainField, DataField otherField,
       {CompareType type = CompareType.equals}) {
-    return JoinedQuerySource(
+    return TripleJoinQuerySource(
       main,
-      joins.followedBy([BeanJoin(other, mainField, type, otherField)]).toList(),
+      joinB,
+      BeanJoin<TDao>(other, mainField, type, otherField),
+    );
+  }
+}
+
+class TripleJoinQuerySource<Ta, Tb, Tc>
+    extends QuerySource<Triple<Ta, Tb, Tc>> {
+  final DataBean<Ta> main;
+  final BeanJoin<Tb> joinB;
+  final BeanJoin<Tc> joinC;
+
+  List<BeanJoin> get joins => [joinB, joinC];
+
+  TripleJoinQuerySource(this.main, this.joinB, this.joinC);
+
+  @override
+  Triple<Ta, Tb, Tc> map(List<QueryResult> results) {
+    return Triple(
+      main.map(results),
+      joinB.bean.map(results),
+      joinC.bean.map(results),
     );
   }
 }